api/requirements.txt
tensorflow==2.7.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.7.0
matplotlib
numpy